+++ ee/tabby-ui/components/chat/question-answer.tsx
@@ -1,50 +1,32 @@
 // Inspired by Chatbot-UI and modified to fit the needs of this project
 // @see https://github.com/mckaywrigley/chatbot-ui/blob/main/components/Chat/ChatMessage.tsx
 
-import React, { forwardRef, useState } from 'react'
+import React, { useMemo } from 'react'
 import Image from 'next/image'
 import tabbyLogo from '@/assets/tabby.png'
-import { compact, isNil } from 'lodash-es'
-import remarkGfm from 'remark-gfm'
-import remarkMath from 'remark-math'
+import { compact, concat, isNil } from 'lodash-es'
 import type { Context } from 'tabby-chat-panel'
 
+import { MessageAttachmentCode } from '@/lib/gql/generates/graphql'
 import { useMe } from '@/lib/hooks/use-me'
 import { filename2prism } from '@/lib/language-utils'
 import {
   AssistantMessage,
+  AttachmentCodeItem,
   QuestionAnswerPair,
-  RelevantCodeContext,
   UserMessage
 } from '@/lib/types/chat'
-import { cn } from '@/lib/utils'
-import { CodeBlock } from '@/components/ui/codeblock'
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger
-} from '@/components/ui/tooltip'
-import { MemoizedReactMarkdown } from '@/components/markdown'
+import { cn, formatLineHashForCodeBrowser } from '@/lib/utils'
 
 import { CopyButton } from '../copy-button'
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger
-} from '../ui/accordion'
+import { ErrorMessageBlock, MessageMarkdown } from '../message-markdown'
 import { Button } from '../ui/button'
-import {
-  IconExternalLink,
-  IconFile,
-  IconRefresh,
-  IconTrash,
-  IconUser
-} from '../ui/icons'
+import { IconFile, IconRefresh, IconTrash, IconUser } from '../ui/icons'
 import { Separator } from '../ui/separator'
 import { Skeleton } from '../ui/skeleton'
 import { UserAvatar } from '../user-avatar'
 import { ChatContext } from './chat'
+import { CodeReferences } from './code-references'
 
 interface QuestionAnswerListProps {
   messages: QuestionAnswerPair[]
@@ -65,6 +47,7 @@ function QuestionAnswerList({
             <QuestionAnswerItem
               isLoading={isLastItem ? isLoading : false}
               message={message}
+              isLastItem={isLastItem}
             />
             {!isLastItem && <Separator className="my-4 md:my-8" />}
           </React.Fragment>
@@ -77,6 +60,7 @@ function QuestionAnswerList({
 interface QuestionAnswerItemProps {
   message: QuestionAnswerPair
   isLoading: boolean
+  isLastItem?: boolean
 }
 
 type SelectCode = {
@@ -84,7 +68,11 @@ type SelectCode = {
   isMultiLine: boolean
 }
 
-function QuestionAnswerItem({ message, isLoading }: QuestionAnswerItemProps) {
+function QuestionAnswerItem({
+  message,
+  isLoading,
+  isLastItem
+}: QuestionAnswerItemProps) {
   const { user, assistant } = message
 
   return (
@@ -98,6 +86,7 @@ function QuestionAnswerItem({ message, isLoading }: QuestionAnswerItemProps) {
             userMessage={user}
             isLoading={isLoading}
             userMessageId={user.id}
+            enableRegenerating={isLastItem}
           />
         </>
       )}
@@ -202,17 +191,19 @@ function UserMessageCard(props: { message: UserMessage }) {
 
 function UserMessageCardActions(props: { message: UserMessage }) {
   const { message } = props
-  const { handleMessageAction } = React.useContext(ChatContext)
+  const { handleMessageAction, isLoading } = React.useContext(ChatContext)
   return (
     <ChatMessageActionsWrapper>
-      <Button
-        variant="ghost"
-        size="icon"
-        onClick={e => handleMessageAction?.(message.id, 'delete')}
-      >
-        <IconTrash />
-        <span className="sr-only">Delete message</span>
-      </Button>
+      {!isLoading && (
+        <Button
+          variant="ghost"
+          size="icon"
+          onClick={e => handleMessageAction?.(message.id, 'delete')}
+        >
+          <IconTrash />
+          <span className="sr-only">Delete message</span>
+        </Button>
+      )}
     </ChatMessageActionsWrapper>
   )
 }
@@ -222,17 +213,29 @@ interface AssistantMessageCardProps {
   isLoading: boolean
   message: AssistantMessage
   userMessage: UserMessage
+  enableRegenerating?: boolean
 }
 
 interface AssistantMessageActionProps {
   userMessageId: string
   message: AssistantMessage
+  enableRegenerating?: boolean
 }
 
 function AssistantMessageCard(props: AssistantMessageCardProps) {
-  const { message, userMessage, isLoading, userMessageId, ...rest } = props
-  const { onNavigateToContext, client } = React.useContext(ChatContext)
-  const contexts: Array<Context> = React.useMemo(() => {
+  const {
+    message,
+    userMessage,
+    isLoading,
+    userMessageId,
+    enableRegenerating,
+    ...rest
+  } = props
+  const { onNavigateToContext, client, onApplyInEditor, onCopyContent } =
+    React.useContext(ChatContext)
+  const [relevantCodeHighlightIndex, setRelevantCodeHighlightIndex] =
+    React.useState<number | undefined>(undefined)
+  const serverCode: Array<Context> = React.useMemo(() => {
     return (
       message?.relevant_code?.map(code => {
         const start_line = code?.startLine ?? 0
@@ -253,12 +256,58 @@ function AssistantMessageCard(props: AssistantMessageCardProps) {
     )
   }, [message?.relevant_code])
 
-  const clientContexts: Array<Context> = React.useMemo(() => {
+  const clientCode: Array<Context> = React.useMemo(() => {
     return compact([
       userMessage.activeContext,
       ...(userMessage?.relevantContext ?? [])
     ])
   }, [userMessage.activeContext, userMessage.relevantContext])
+
+  const attachmentCode: Array<AttachmentCodeItem> = useMemo(() => {
+    return concat(clientCode, serverCode).map(o => ({
+      content: o.content,
+      filepath: o.filepath,
+      gitUrl: o.git_url,
+      startLine: o.range.start,
+      // FIXME
+      language: ''
+    }))
+  }, [clientCode, serverCode])
+
+  const onCodeCitationMouseEnter = (index: number) => {
+    setRelevantCodeHighlightIndex(index - 1 - (attachmentCode?.length || 0))
+  }
+
+  const onCodeCitationMouseLeave = (index: number) => {
+    setRelevantCodeHighlightIndex(undefined)
+  }
+
+  const openCodeBrowserTab = (code: MessageAttachmentCode) => {
+    const start_line = code?.startLine ?? 0
+    const lineCount = code.content.split('\n').length
+    const end_line = start_line + lineCount - 1
+
+    if (!code.filepath) return
+    const url = new URL(`${window.location.origin}/files`)
+    const searchParams = new URLSearchParams()
+    searchParams.append('redirect_filepath', code.filepath)
+    searchParams.append('redirect_git_url', code.gitUrl)
+    url.search = searchParams.toString()
+
+    const lineHash = formatLineHashForCodeBrowser({
+      start: start_line,
+      end: end_line
+    })
+    if (lineHash) {
+      url.hash = lineHash
+    }
+
+    window.open(url.toString())
+  }
+
+  const onCodeCitationClick = (code: MessageAttachmentCode) => {
+    openCodeBrowserTab(code)
+  }
 
   return (
     <div
@@ -279,26 +328,36 @@ function AssistantMessageCard(props: AssistantMessageCardProps) {
           <AssistantMessageCardActions
             message={message}
             userMessageId={userMessageId}
+            enableRegenerating={enableRegenerating}
           />
         </div>
       </div>
 
       <div className="w-full flex-1 space-y-2 overflow-hidden px-1 md:ml-4">
         <CodeReferences
-          contexts={contexts}
-          userContexts={clientContexts}
+          contexts={serverCode}
+          userContexts={clientCode}
           onContextClick={(ctx, isInWorkspace) => {
             onNavigateToContext?.(ctx, {
               openInEditor: client === 'vscode' && isInWorkspace
             })
           }}
           isExternalLink={!!client}
+          highlightIndex={relevantCodeHighlightIndex}
         />
         {isLoading && !message?.message ? (
           <MessagePendingIndicator />
         ) : (
           <>
-            <MessageMarkdown message={message.message} />
+            <MessageMarkdown
+              message={message.message}
+              onApplyInEditor={onApplyInEditor}
+              onCopyContent={onCopyContent}
+              attachmentCode={attachmentCode}
+              onCodeCitationClick={onCodeCitationClick}
+              onCodeCitationMouseEnter={onCodeCitationMouseEnter}
+              onCodeCitationMouseLeave={onCodeCitationMouseLeave}
+            />
             {!!message.error && <ErrorMessageBlock error={message.error} />}
           </>
         )}
@@ -319,10 +378,10 @@ function AssistantMessageCardActions(props: AssistantMessageActionProps) {
     isLoading: isGenerating,
     onCopyContent
   } = React.useContext(ChatContext)
-  const { message, userMessageId } = props
+  const { message, userMessageId, enableRegenerating } = props
   return (
     <ChatMessageActionsWrapper>
-      {!isGenerating && (
+      {!isGenerating && enableRegenerating && (
         <Button
           variant="ghost"
           size="icon"
@@ -334,93 +393,6 @@ function AssistantMessageCardActions(props: AssistantMessageActionProps) {
       )}
       <CopyButton value={message.message} onCopyContent={onCopyContent} />
     </ChatMessageActionsWrapper>
-  )
-}
-
-function MessageMarkdown({ message }: { message: string }) {
-  const { onCopyContent, onApplyInEditor } = React.useContext(ChatContext)
-  return (
-    <MemoizedReactMarkdown
-      className="prose max-w-none break-words dark:prose-invert prose-p:leading-relaxed prose-pre:mt-1 prose-pre:p-0"
-      remarkPlugins={[remarkGfm, remarkMath]}
-      components={{
-        p({ children }) {
-          return <p className="mb-2 last:mb-0">{children}</p>
-        },
-        code({ node, inline, className, children, ...props }) {
-          if (children.length) {
-            if (children[0] == '▍') {
-              return (
-                <span className="mt-1 animate-pulse cursor-default">▍</span>
-              )
-            }
-
-            children[0] = (children[0] as string).replace('`▍`', '▍')
-          }
-
-          const match = /language-(\w+)/.exec(className || '')
-          const metaObject = parseMetaDataString(node.data?.meta as string)
-          const isReference = metaObject?.['is_reference'] === '1'
-
-          if (isReference) {
-            return null
-          }
-
-          if (inline) {
-            return (
-              <code className={className} {...props}>
-                {children}
-              </code>
-            )
-          }
-
-          return (
-            <CodeBlock
-              key={Math.random()}
-              language={(match && match[1]) || ''}
-              value={String(children).replace(/\n$/, '')}
-              onCopyContent={onCopyContent}
-              onApplyInEditor={onApplyInEditor}
-              {...props}
-            />
-          )
-        }
-      }}
-    >
-      {message}
-    </MemoizedReactMarkdown>
-  )
-}
-
-function ErrorMessageBlock({ error = 'Fail to fetch' }: { error?: string }) {
-  const errorMessage = React.useMemo(() => {
-    let jsonString = JSON.stringify(
-      {
-        error: true,
-        message: error
-      },
-      null,
-      2
-    )
-    const markdownJson = '```\n' + jsonString + '\n```'
-    return markdownJson
-  }, [error])
-  return (
-    <MemoizedReactMarkdown
-      className="prose break-words text-sm dark:prose-invert prose-p:leading-relaxed prose-pre:mt-1 prose-pre:p-0"
-      remarkPlugins={[remarkGfm, remarkMath]}
-      components={{
-        code({ node, inline, className, children, ...props }) {
-          return (
-            <div {...props} className={cn(className, 'bg-zinc-950 p-2')}>
-              {children}
-            </div>
-          )
-        }
-      }}
-    >
-      {errorMessage}
-    </MemoizedReactMarkdown>
   )
 }
 
@@ -457,197 +429,6 @@ function ChatMessageActionsWrapper({
       {...props}
     />
   )
-}
-
-interface ContextReferencesProps {
-  contexts: RelevantCodeContext[]
-  userContexts?: RelevantCodeContext[]
-  className?: string
-  onContextClick?: (
-    context: RelevantCodeContext,
-    isInWorkspace?: boolean
-  ) => void
-  defaultOpen?: boolean
-  enableTooltip?: boolean
-  onTooltipClick?: () => void
-  isExternalLink?: boolean
-  highlightIndex?: number | undefined
-}
-
-export const CodeReferences = forwardRef<
-  HTMLDivElement,
-  ContextReferencesProps
->(
-  (
-    {
-      contexts,
-      userContexts,
-      className,
-      onContextClick,
-      defaultOpen,
-      enableTooltip,
-      onTooltipClick,
-      isExternalLink,
-      highlightIndex
-    },
-    ref
-  ) => {
-    const totalContextLength = (userContexts?.length || 0) + contexts.length
-    const isMultipleReferences = totalContextLength > 1
-
-    if (totalContextLength === 0) return null
-    return (
-      <Accordion
-        type="single"
-        collapsible
-        className={cn('bg-transparent text-foreground', className)}
-        defaultValue={defaultOpen ? 'references' : undefined}
-        ref={ref}
-      >
-        <AccordionItem value="references" className="my-0 border-0">
-          <AccordionTrigger className="my-0 py-2 font-semibold">
-            <span className="mr-2">{`Read ${totalContextLength} file${
-              isMultipleReferences ? 's' : ''
-            }`}</span>
-          </AccordionTrigger>
-          <AccordionContent className="space-y-2">
-            {userContexts?.map((item, index) => {
-              return (
-                <ContextItem
-                  key={`user-${index}`}
-                  context={item}
-                  onContextClick={ctx => onContextClick?.(ctx, true)}
-                />
-              )
-            })}
-            {contexts.map((item, index) => {
-              return (
-                <ContextItem
-                  key={`assistant-${index}`}
-                  context={item}
-                  onContextClick={ctx => onContextClick?.(ctx, false)}
-                  enableTooltip={enableTooltip}
-                  onTooltipClick={onTooltipClick}
-                  isExternalLink={isExternalLink}
-                  isHighlighted={highlightIndex === index}
-                />
-              )
-            })}
-          </AccordionContent>
-        </AccordionItem>
-      </Accordion>
-    )
-  }
-)
-CodeReferences.displayName = 'CodeReferences'
-
-function ContextItem({
-  context,
-  clickable = true,
-  onContextClick,
-  enableTooltip,
-  onTooltipClick,
-  isExternalLink,
-  isHighlighted
-}: {
-  context: RelevantCodeContext
-  clickable?: boolean
-  onContextClick?: (context: RelevantCodeContext) => void
-  enableTooltip?: boolean
-  onTooltipClick?: () => void
-  isExternalLink?: boolean
-  isHighlighted?: boolean
-}) {
-  const [tooltipOpen, setTooltipOpen] = useState(false)
-  const isMultiLine =
-    !isNil(context.range?.start) &&
-    !isNil(context.range?.end) &&
-    context.range.start < context.range.end
-  const pathSegments = context.filepath.split('/')
-  const fileName = pathSegments[pathSegments.length - 1]
-  const path = pathSegments.slice(0, pathSegments.length - 1).join('/')
-  const scores = context?.extra?.scores
-  const onTooltipOpenChange = (v: boolean) => {
-    if (!enableTooltip || !scores) return
-
-    setTooltipOpen(v)
-  }
-
-  return (
-    <Tooltip
-      open={tooltipOpen}
-      onOpenChange={onTooltipOpenChange}
-      delayDuration={0}
-    >
-      <TooltipTrigger asChild>
-        <div
-          className={cn('rounded-md border p-2', {
-            'cursor-pointer hover:bg-accent': clickable,
-            'cursor-default pointer-events-auto': !clickable,
-            'bg-accent transition-all': isHighlighted
-          })}
-          onClick={e => clickable && onContextClick?.(context)}
-        >
-          <div className="flex items-center gap-1 overflow-hidden">
-            <IconFile className="shrink-0" />
-            <div className="flex-1 truncate" title={context.filepath}>
-              <span>{fileName}</span>
-              {context.range?.start && (
-                <span className="text-muted-foreground">
-                  :{context.range.start}
-                </span>
-              )}
-              {isMultiLine && (
-                <span className="text-muted-foreground">
-                  -{context.range.end}
-                </span>
-              )}
-              <span className="ml-2 text-xs text-muted-foreground">{path}</span>
-            </div>
-            {isExternalLink && (
-              <IconExternalLink className="shrink-0 text-muted-foreground" />
-            )}
-          </div>
-        </div>
-      </TooltipTrigger>
-      <TooltipContent
-        align="start"
-        onClick={onTooltipClick}
-        className="cursor-pointer p-2"
-      >
-        <div>
-          <div className="mb-2 font-semibold">Scores</div>
-          <div className="space-y-1">
-            <div className="flex">
-              <span className="w-20">rrf:</span>
-              {scores?.rrf ?? '-'}
-            </div>
-            <div className="flex">
-              <span className="w-20">bm25:</span>
-              {scores?.bm25 ?? '-'}
-            </div>
-            <div className="flex">
-              <span className="w-20">embedding:</span>
-              {scores?.embedding ?? '-'}
-            </div>
-          </div>
-        </div>
-      </TooltipContent>
-    </Tooltip>
-  )
-}
-
-function parseMetaDataString(metaData: string | undefined) {
-  const metadataObj: Record<string, string> = {}
-  if (!metaData) return metadataObj
-
-  const keyValuePairs = metaData.split(' ')
-  keyValuePairs.forEach(pair => {
-    const [key, value] = pair.split('=')
-    metadataObj[key] = value
-  })
-
-  return metadataObj
 }
 
 export { QuestionAnswerList }
